Update README.md (#17)
* Update README.md

I've updated to follow the guidelines I've previously received from Microsoft marketing on how to refer to Microsoft SQL Server in various scenarios.  In short, the long name is Microsoft SQL Server, the short name is SQL Server.  It's fine to use MSSQL in software as an abbreviation for file and directory names, configuration variables, logs,  but it's not on brand to refer to the product in documentation/text as MSSQL.  I also saw that the EULA acceptance for SQL Server was listed as being for mssql-conf.  It's actually for more than that - that's the general SQL Server software EULA they're accepting.

* Fix MSSQL in meta/main.yml, fix typo

Co-authored-by: Sergei Petrosian <spetrosi@redhat.com>