Remove the workaround for `any_errors_fatal: true` (#100)
Previously, the role would use a workaround of setting a
__mssql_primary_successful variable to define whether the primary node
really failed or not. This was required because any_errors_fatal: true
would not work when run in a block where rescue or always block would
execute.
This workaround creates troubles when the order of hosts in inventory is
not primary-secondary-witness - the role skips a rescue block after
failing tasks in the block on primary and goes to the next block.
Removing the workardound for stability. Certificate and private key now
won't be removed from the control node in the case of failure, but those
files will be removed on a next successful role invocation.