ansible-lint 6.x fixes
The big one is that ansible-lint doesn't like templates in `name`
strings except at the end.  In general, Ansible does not like having
templated variables in `name` values because it makes it harder to
grep the source to find a log message in the source.
The other ones are jinja spacing cleanup, use of `true`/`false`
instead of `yes`/`no`, and various other cleanup.