ci: Configure Packit propose_downstream job (#293)
When there is a new release of this repo, Packit will fetch
linux-system-roles.spec from the Rawhide branch in Fedora dist-git,
update it (Source lines and %setup will be regenerated) and open a PR in
Fedora Pagure. It will also upload new tarballs to lookaside and update
sources.

The PR will be opened only against Rawhide for now, can be changed in
dist_git_branches.

documentation:
https://packit.dev/docs/configuration/#propose_downstream
https://packit.dev/docs/fedora-releases-guide/

Taking inspiration from
https://github.com/linux-system-roles/auto-maintenance/pull/233