fix: Skip runtime operations in non-systemd environments
Cause: Both the actual role and the tests were previously assuming that
the system was booted with systemd and could run services. But this is
not the case when running the role during container builds.

Consequence: The role did not work during bootc container builds.

Fix: Detect if the system is booted (with systemd), and skip all runtime
operations and checks if not.

Result: The role now works during container builds.

Setting the tuned profile is a bit complicated: Reading the current
profile can happen via the config file (which is even easier than having
to manipulate the string from `tuned-adm active`). But the profile
changing has a workaround for rhbz#1825882 which still applies to RHEL
8.{0,2,4}. These aren't supported for bootc, however. So keep two
parallel code paths for these. Once the hack can be dropped, it is
enough to use the file based approach and restart tuned.service.

Verifying the mssql edition is also a nuisance: Conceptually, running
`/opt/mssql/bin/sqlservr -v` is better than scraping the edition out of
the log file, but unfortunately that command fails if the server is
already running. So we need this new code path for the non-booted case,
and keep checking the log file for the traditional booted case. But
clean that up a bit to be structurally similar.