tests: Fix running non-SELinux tests after SELinux ones
Running a test which enables SELinux confinement (like
tests_input_sql_file_2022) before a test which doesn't (like
tests_configure_ha_cluster_external) messed up the latter: After the
first test, the `mssql` SELinux module was still installed and active,
even though the m-s-selinux RPM was uninstalled. This broke permissions
when the unconfined server started up, as the custom directory contexts
were still in place.

The m-s-selinux post-uninstall script tries to remove the module, but it
(silently) fails on

> libsepol.context_from_record: type mssql_var_t is not defined (No such file or directory).

Fix this by dropping the local customizations first, and re-remove the
module.

However, even after this there is *still* some residue. Invoking the
`selinux` role in a situation where m-s-selinux.rpm is not installed
tries to apply the nonexisting `mssql_db_t` type to /tmp/mssql_log,
which fails:

> Type mssql_db_t is invalid, must be a file or device type

This could possibly be hacked with some seddery in /etc/selinux/, but
let's not do that -- we don't actually need to reset the SELinux policy
if we didn't touch it in the first place, so make this cleanup step
conditional.