tests: Add workaround for Ansible 2.9 SELinux cleanup failure
The "conditionalize SELinux cleanup" hack in commit e1c8d1fdf973f that
fixed a regression with different test order caused a regression itself:
On older Ansible (observed on 2.9) it fails some tests with:

> The conditional check 'mssql_manage_selinux | d(false) | bool' failed.
> The error was: error while evaluating conditional
> (mssql_manage_selinux | d(false) | bool): {{ mssql_run_selinux_confined }}:
> 'mssql_run_selinux_confined' is undefined

So the `d()` default does not apply to the actual definition of the
variable in these versions.

Avoid that by adding a default to ensure that the value isn't undefined.