tests: Update tests_password_2022 to do bootc end-to-end validation
This only covers the initial role invocation. The others which change
passwords and ports are skipped during the validation run. In principle
it would be interesting to run them on the deployed QEMU side, but in
practice that currently fails with

> Configure to run as a confined application with SELinux
> tests/roles/linux-system-roles.mssql/tasks/main.yml:312
> missing required arguments: backend, image

due to the partial invocation and hence missing global variables. So
let's defer that until we do a concerted effort of running the roles on
bootc deployments.

See https://issues.redhat.com/browse/RHEL-78157